requirements.txt
streamlit
requests
pandas
streamlit-autorefresh
fastapi
uvicorn[standard]
python-dotenv